chore(meshes): makes MeshDetails injectable
